refer sample

+++ app/api/item/delete/[id]/route.js
@@ -3,13 +3,18 @@ import connectDB from "../../../../utils/database"
 import { ItemModel } from "../../../../utils/schemaModels"
 
 export async function DELETE(request, context){
-    const params = await context.params
+    const reqBody = await request.json()
     try{
         await connectDB()
-        await ItemModel.deleteOne({_id: params.id})
-        return NextResponse.json({message: "アイテム削除成功"})
+        const params = await context.params 
+        const singleItem = await ItemModel.findById(params.id)
+        if(singleItem.email === reqBody.email){
+            await ItemModel.deleteOne({_id: params.id})
+            return NextResponse.json({message: "アイテム削除成功"})
+        }else{
+            return NextResponse.json({message: "他の人が作成したアイテムです"})
+        }
     }catch(err){
-        return NextResponse.json({message: "アイテム削除に失敗しました"})
-    }  
+        return NextResponse.json({message: "アイテム削除失敗"})
+    }
 }
-
